Reinstate blur code but add time delay to the run.later
The code commented out in a previous commit was necessary to remove the search list when the search bar was blurred.

There was some kind of race condition whereby the blur was recognized before the clickNewWord action could process.

The Ember.run.later function allows a time to be specified. Adding a 300ms delay allows the clickNewWord action to process before the blur is registered.