Should not redirect to registration page when signing in
If a user goes from register to sign in page (either through our own redirect or via the link on the registration page), their previous route will be users.new.

We do not want people to be redirected to "users.new" if they just sign in. Therefore, I added a catch for the users.new route – if that is the previous route, we redirect to the index instead.

(Ideally, it would redirect to the page before the registration page but I don't quite know how to do that...)