Finished updating FAQ
Added new FAQ questions and answers. Changed WordSet spelling to Wordset a couple places. Re organized the order of some questions. Took out old how proposals work section and replaced it with a new and improved version of what was once in the guidlines. Changed talk of editors or users to "members."